Import a simple Script (like mysqlDump, or from Sql server)
Only tables name,columns and FK are imported (as a class/attribute/
association or generalization). No stereotype yet.
Ignore DROP TABLE, for testing some MySQL script get with
 mysqldump -u root -p --no-data icescrum > myFile.sql
Setting for choosing dump Database (MySQL / SqlServer)
Setting for building model :
 If a column belong to a FK, enable to create it or not in the model

Initial value for Mysql and SqlServer (String / Numeric)
If a is value is null, multiplicity is 0..1, otherwise 1.
For a FK, if a column can be null then I choose that the association is 
0..1

Contributed-by: Laurent BRAUD

Change-Id: I1d0d67f67559aba28a43535b8e913245bf70108b

git-svn-id: http://argouml-sql.tigris.org/svn/argouml-sql/trunk@306 dc9a2ffb-671e-0410-bcaf-86adf31cdc5b